chore: Introduce common directory to deduplicate shared code
