+++ common/sysconfig.cpp
@@ -1,6 +1,5 @@
 /*  Copyright 2024 Pretendo Network contributors <pretendo.network>
     Copyright 2024 Ash Logan <ash@heyquark.com>
-    Copyright 2023 Maschell
     Copyright 2020-2022 V10lator <v10lator@myway.de>
     Copyright 2022 Xpl0itU <DaThinkingChair@protonmail.com>
 
@@ -81,8 +80,8 @@ static void get_mcp_config() {
     mcp_os_version = os_version;
 
     DEBUG_FUNCTION_LINE_VERBOSE("Running on %d.%d.%d%c; %s%s",
-        os_version.major, os_version.minor, os_version.patch, os_version.region
-        config.code_id, config.serial_id
+                                os_version.major, os_version.minor, os_version.patch, os_version.region
+                                        config.code_id, config.serial_id
     );
 }
 
@@ -96,4 +95,21 @@ MCPSystemVersion get_console_os_version() {
     if (!mcp_os_version) get_mcp_config();
 
     return mcp_os_version.value_or((MCPSystemVersion) { .major = 5, .minor = 5, .patch = 5, .region = 'E' });
+}
+
+static inline int digit(char a) {
+    if (a < '0' || a > '9') return 0;
+    return a - '0';
+}
+
+unsigned short get_console_peertopeer_port() {
+    const char * serial = get_console_serial();
+
+    unsigned short port = 50000 +
+                          (digit(serial[4]) * 1000) +
+                          (digit(serial[5]) * 100 ) +
+                          (digit(serial[6]) * 10  ) +
+                          (digit(serial[7]) * 1   );
+
+    return port;
 }
