+++ common/sysconfig.h
@@ -11,5 +11,6 @@
 nn::swkbd::LanguageType get_system_language();
 const char * get_console_serial();
 MCPSystemVersion get_console_os_version();
+unsigned short get_console_peertopeer_port();
 
 #endif //INKAY_SYSCONFIG_H
