--- plugin/src/Notification.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void ShowNotification(const char* notification);
