--- plugin/src/utils/sysconfig.h
@@ -1,12 +1,0 @@
-//
-// Created by ash on 9/04/24.
-//
-
-#ifndef INKAY_SYSCONFIG_H
-#define INKAY_SYSCONFIG_H
-
-#include <nn/swkbd.h>
-
-nn::swkbd::LanguageType get_system_language();
-
-#endif //INKAY_SYSCONFIG_H
