+++ src/config.cpp
@@ -20,44 +20,3 @@
 bool Config::connect_to_network = false;
 bool Config::initialized = false;
 bool Config::shown_uninitialized_warning = false;
-
-static config_strings strings;
-
-config_strings get_config_strings(nn::swkbd::LanguageType language) {
-    switch (language) {
-        case nn::swkbd::LanguageType::English:
-        default: return {
-#include "lang/en_US.lang"
-        };
-        case nn::swkbd::LanguageType::Spanish: return {
-#include "lang/es_ES.lang"
-        };
-        case nn::swkbd::LanguageType::French: return {
-#include "lang/fr_FR.lang"
-        };
-        case nn::swkbd::LanguageType::Italian: return {
-#include "lang/it_IT.lang"
-        };
-        case nn::swkbd::LanguageType::German: return {
-#include "lang/de_DE.lang"
-        };
-        case nn::swkbd::LanguageType::SimplifiedChinese: return {
-#include "lang/zh_CN.lang"
-        };
-        case nn::swkbd::LanguageType::TraditionalChinese: return {
-#include "lang/zh_Hant.lang"
-        };
-        case nn::swkbd::LanguageType::Portuguese: return {
-#include "lang/pt_BR.lang"
-        };
-        case nn::swkbd::LanguageType::Japanese: return {
-#include "lang/ja_JP.lang"
-        };
-        case nn::swkbd::LanguageType::Dutch: return {
-#include "lang/nl_NL.lang"
-        };
-        case nn::swkbd::LanguageType::Russian: return {
-#include "lang/ru_RU.lang"
-        };
-    }
-}
